back-end/requirements.txt
alembic==1.0.2
blinker==1.4
certifi==2018.10.15
chardet==3.0.4
Click==7.0
colorama==0.4.0
coverage==4.5.2
Flask==1.0.2
Flask-Cors==3.0.6
Flask-HTTPAuth==3.2.4
Flask-Mail==0.9.1
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.3.2
httpie==0.9.9
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
Pygments==2.2.0
PyJWT==1.6.4
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
requests==2.20.0
six==1.11.0
SQLAlchemy==1.2.13
urllib3==1.24
Werkzeug==0.14.1
